Defer segmentation and UDP/TCP checksumming
...to when the outgoing NIC is guaranteed as the outgoing NIC for a
packet may change as the packet flows down the stack.

Example, if NAT is performed on a packet which was originally to be
sent out through NIC A but gets rerouted to be sent out NIC B but
the NICs have different capabilities (NIC B does not support HW GSO
or checksum offloading), the packet may never be accepted by the
remote as it will be malformed (when checksum offload is not possible)
or too big to be transmitted (when HW GSO is not available).

This change has the added benefit of decoupling the transport endpoints
from outgoing NICs so that outgoing NICs may change underneath the
endpoint without creating unusable/stale endpoints.

PiperOrigin-RevId: 364938510